Add comment about websocket token in README
Move tests to clean up redundant module
