+++ p2poolv2_lib/src/store/dag_store.rs
@@ -2338,4 +2338,77 @@ mod tests {
             "chain tip must never appear as an uncle"
         );
     }
+
+    #[test]
+    fn test_share_info_serialization() {
+        let share_info = ShareInfo {
+            blockhash: BlockHash::all_zeros(),
+            prev_blockhash: BlockHash::all_zeros(),
+            height: 42,
+            miner_pubkey: "02aabbccdd".to_string(),
+            timestamp: 1_700_000_000,
+            bits: CompactTarget::from_consensus(0x1b4188f5),
+            uncles: vec![],
+        };
+
+        let json = serde_json::to_string(&share_info).unwrap();
+        assert!(json.contains("\"height\":42"));
+        assert!(json.contains("\"miner_pubkey\":\"02aabbccdd\""));
+        assert!(json.contains("\"timestamp\":1700000000"));
+    }
+
+    #[test]
+    fn test_share_info_with_uncles_serialization() {
+        let uncle = UncleInfo {
+            blockhash: BlockHash::all_zeros(),
+            prev_blockhash: BlockHash::all_zeros(),
+            miner_pubkey: "02uncle".to_string(),
+            timestamp: 1_700_000_010,
+            height: Some(41),
+        };
+
+        let share_info = ShareInfo {
+            blockhash: BlockHash::all_zeros(),
+            prev_blockhash: BlockHash::all_zeros(),
+            height: 42,
+            miner_pubkey: "02parent".to_string(),
+            timestamp: 1_700_000_020,
+            bits: CompactTarget::from_consensus(0x1b4188f5),
+            uncles: vec![uncle],
+        };
+
+        let json = serde_json::to_string(&share_info).unwrap();
+        assert!(json.contains("\"02uncle\""));
+        assert!(json.contains("\"height\":41"));
+    }
+
+    #[test]
+    fn test_uncle_info_serialization() {
+        let uncle = UncleInfo {
+            blockhash: BlockHash::all_zeros(),
+            prev_blockhash: BlockHash::all_zeros(),
+            miner_pubkey: "02aabb".to_string(),
+            timestamp: 1_700_000_005,
+            height: Some(10),
+        };
+
+        let json = serde_json::to_string(&uncle).unwrap();
+        assert!(json.contains("\"miner_pubkey\":\"02aabb\""));
+        assert!(json.contains("\"timestamp\":1700000005"));
+        assert!(json.contains("\"height\":10"));
+    }
+
+    #[test]
+    fn test_uncle_info_with_no_height_serialization() {
+        let uncle = UncleInfo {
+            blockhash: BlockHash::all_zeros(),
+            prev_blockhash: BlockHash::all_zeros(),
+            miner_pubkey: "02ccdd".to_string(),
+            timestamp: 1_700_000_005,
+            height: None,
+        };
+
+        let json = serde_json::to_string(&uncle).unwrap();
+        assert!(json.contains("\"height\":null"));
+    }
 }
